requirements.txt
flask
torch 
torchvision 
pillow
tensorflow
keras
numpy
pandas
